No ticket (follow-up of KOMPVZ-120 ): When start listening on 'ready' event of Cadenza, we log the origin and the window we're listening on. We log the same things on Cadenza's side now. This shall make it easier to detect a mismatch between both which causes PostMessages to not be received.
No ticket (follow-up of KOMPVZ-120 ): When start listening on 'ready' event of Cadenza, we log the origin and the window we're listening on. We log the same things on Cadenza's side now. This shall make it easier to detect a mismatch between both which causes PostMessages to not be received.